+++ test/test-node-api.js
@@ -4,7 +4,7 @@ const path = require('path');
 const { fileURLToPath } = require('url');
 const EC = require('eight-colors');
 
-const CoverageReport = require('../');
+const MCR = require('../');
 
 // test lib app
 const {
@@ -37,7 +37,7 @@ const coverageOptions = {
 
 const generate = async () => {
 
-    const coverageReport = new CoverageReport(coverageOptions);
+    const coverageReport = MCR(coverageOptions);
 
     // clean cache before add coverage data
     coverageReport.cleanCache();