+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Changelog
 
+* 2.1.0 (Breaking Change) 
+    - defaults to export `createCoverageReport` function instead of `CoverageReport` class
+    ```diff
+    - import CoverageReport, { createCoverageReport } from 'monocart-coverage-reports';
+    + import createCoverageReport, { CoverageReport } from 'monocart-coverage-reports';
+    ```
+
 * 2.0.10
     - updated dependencies